pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.135.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
    "starlette>=0.52.1",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
